+++ ConcernsCaseWork/ConcernsCaseWork.API.Tests/Factories/FinancialPlanFactoryTests.cs
@@ -1,6 +1,5 @@
 ﻿using ConcernsCaseWork.API.Contracts.FinancialPlan;
 using ConcernsCaseWork.API.Factories.CaseActionFactories;
-using ConcernsCaseWork.API.ResponseModels.CaseActions.FinancialPlan;
 using ConcernsCaseWork.Data.Models;
 using FizzWare.NBuilder;
 using FluentAssertions;
